refactored form out of shapeshift into OrderForm component
