+++ app/components/ShapeShift.js
@@ -1,24 +1,16 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import axios from "axios";
-import numeral from "numeral";
-import { Link } from "react-router";
-
-import { setMarketPrice, resetPrice } from "../modules/wallet";
-import { togglePane } from "../modules/dashboard";
 import { initiateGetBalance, intervals } from "../components/NetworkSwitch";
+
 import UnavailableExchange from "../components/UnavailableExchange";
+import OrderForm from "../components/OrderForm";
+import Order from "./Order";
 import Deposit from "../components/Deposit";
 import Process from "../components/Process";
 import Complete from "../components/Complete";
+
 import { sendEvent, clearTransactionEvent } from "../modules/transactions";
 import {fetchNeoStatus, startShiftOrder, fetchDepositStatus, resetOrderState} from "../modules/shapeshift";
-
-import neoLogo from "../img/neo.png";
-import NeoLogo from "./Brand/Neo";
-import BtcLogo from "./Brand/Bitcoin";
-import shapeshiftLogo from "../img/shapeshift.png";
-import Order from "./Order";
 
 
 // force sync with balance data
@@ -29,32 +21,13 @@ const refreshBalance = async (dispatch, net, address) => {
 		setTimeout(() => dispatch(clearTransactionEvent()), 1000);
 	});
 };
-const shapeshiftPk = "5aad9888213a9635ecda3ed8bb2dc45c0a8d95dc36da7533c78f3eba8f765ce77538aae79d0e35642e39f208b7428631188f03c930e91f299f9eb40556f8e74d";
 
 
 class ShapeShift extends Component {
 	constructor(props) {
 		super(props);
-		this.state = {
-			gasPrice: 0,
-			rpxPrice: 0,
-			qlcPrice: 0,
-			dbcPrice: 0,
-			selectedAsset: "select",
-			selectedAssetToNeoRate: 0,
-			maxLimit: 0,
-			minLimit: 0,
-			minerFee: 0,
-			depositAmt: 0.000001,
-		};
 		this.pollForNeoConditonallyEvery = this.pollForNeoConditonallyEvery.bind(this);
 		this.pollForDepositStatusConditionallyEvery = this.pollForDepositStatusConditionallyEvery.bind(this);
-		this.getMarketInfo = this.getMarketInfo.bind(this);
-		this.handleSelectAsset = this.handleSelectAsset.bind(this);
-		this.handleDepositAmtChange = this.handleDepositAmtChange.bind(this);
-		this.calcExpectedNeo = this.calcExpectedNeo.bind(this);
-		this.handleOrderClick = this.handleOrderClick.bind(this);
-		this.determineNeoOutputAmtValidity = this.determineNeoOutputAmtValidity.bind(this);
 	}
 
 	componentDidMount() {
@@ -72,7 +45,6 @@ class ShapeShift extends Component {
 		}, ms);
 	}
 
-	// fix this like above
 	pollForDepositStatusConditionallyEvery(ms) {
 		const { fetchDepositStatus, stage, txData } = this.props;
 		if (stage === "depositing" || stage === "processing") fetchDepositStatus(txData.deposit);
@@ -82,199 +54,14 @@ class ShapeShift extends Component {
 		}, ms);
 	}
 
-	async getMarketInfo(asset) {
-		const url = `https://shapeshift.io/marketinfo/${asset}_neo`;
-		const response = await axios.get(url);
-		const { data } = await response;
-		this.setState({
-			selectedAssetToNeoRate: data.rate,
-			maxLimit: data.maxLimit,
-			minLimit: data.minimum,
-			minerFee: data.minerFee
-		});
-	}
-
-	handleSelectAsset(e) {
-		const { value } = e.target;
-		this.getMarketInfo(value);
-		this.setState({ selectedAsset: value });
-	}
-
-	handleDepositAmtChange(e) {
-		const { value } = e.target;
-		this.setState({ depositAmt: value });
-	}
-
-	handleOrderClick() {
-		const { address } = this.props;
-		const { selectedAsset, depositAmt } = this.state;
-		const shiftConfig = {
-			withdrawal: address,
-			pair: `${selectedAsset}_neo`,
-			amount: this.calcExpectedNeo(),
-			returnAddress: "shapeshift_address_here_based_on_selected_asset",
-			apiKey: shapeshiftPk
-		};
-		this.props.startShiftOrder(shiftConfig);
-	}
-
-	calcExpectedNeo() {
-		const { depositAmt, selectedAssetToNeoRate } = this.state;
-		return depositAmt * selectedAssetToNeoRate;
-	}
-
-	determineNeoOutputAmtValidity() {
-		const neoOutput = this.calcExpectedNeo();
-		return parseInt(neoOutput) === neoOutput ? true : false;
-	}
-
-	// Make null stage component
-
 	render() {
 		const { available, fetching, stage, txData, completeData } = this.props;
 		if (!available && !fetching && !stage) return <UnavailableExchange exchangeName={"ShapeShift"}/>;
+		else if (!stage) return <OrderForm {...this.props} />;
 		else if (stage === "ordering") return <Order exchangeName={"ShapeShift"}/>;
 		else if (stage === "depositing") return <Deposit txData={txData} exchangeName={"shapeshift"}/>;
 		else if (stage === "processing") return <Process txData={txData}/>;
 		else if (stage === "complete") return <Complete completeData={completeData}/>;
-		// render component specially for a null stage
-		// render component specially for a ordering stage
-		// render component specially for a complete stage
-
-
-		const isValidNeoOutput = this.determineNeoOutputAmtValidity();
-
-		// Current default rendering is null stage, will render
-		return (
-			<div>
-				<div className="progress-bar fadeInLeft-ex" />
-				<div>Shapeshift stage: {this.props.stage}</div>
-				<div className="row prog-info top-20">
-					<div className="col-xs-2 col-xs-offset-1 sm-text center">
-						Enter Amount to Deposit
-					</div>
-					<div className="col-xs-2 sm-text center grey-out">
-						Placing Your Order</div>
-					<div className="col-xs-2 sm-text center grey-out">
-						Generating Deposit Address
-					</div>
-					<div className="col-xs-2 sm-text center grey-out">
-						Processing Your Order
-					</div>
-					<div className="col-xs-2 sm-text center grey-out">
-						Transaction Complete!
-					</div>
-				</div>
-
-				{!stage && ( <div className="top-130 dash-panel">
-					<h2 className="center">ShapeShift Exchange Service</h2>
-					<hr className="dash-hr-wide" />
-					<div className="row top-10">
-
-						<div className="col-xs-4">
-							<select
-								name="select-profession"
-								id="select-profession"
-								className=""
-								value={this.state.selectedAsset}
-								onChange={this.handleSelectAsset}
-							>
-								<option value="select" disabled={true}>
-									Select Asset
-								</option>
-								<option value="btc">
-									Bitcoin (BTC)
-								</option>
-								<option value="eth">
-									Ethereum (ETH)
-								</option>
-								<option value="ltc">
-									Litecoin (LTC)
-								</option>
-								<option value="xmr">
-									Monero (XMR)
-								</option>
-							</select>
-							<p className="sm-text top-10">
-								Select Asset to Exchange
-							</p>
-						</div>
-
-						<div className="col-xs-4">
-							<input
-								value={this.state.depositAmt}
-								onChange={this.handleDepositAmtChange}
-								className="form-control-exchange center"
-								type="number"
-								min={0.01}
-							/>
-							<p className="sm-text">
-								Amount to Deposit
-							</p>
-						</div>
-
-
-						<div className="col-xs-4">
-							{
-								isValidNeoOutput
-									? <div></div>
-									: <div style={{ color: "red" }}>Sorry, NEO outputs must be whole numbers :(</div>
-							}
-							<input
-								className="form-control-exchange center"
-								value={this.calcExpectedNeo()}
-								placeholder="0"
-								disabled
-							/>
-							<p className="sm-text">
-								Amount of NEO Received
-							</p>
-						</div>
-					</div>
-
-					<div className="row">
-						<div className="col-xs-8 top-20">
-
-							<input
-								className="form-control-exchange center"
-								disabled
-								placeholder={this.props.address}
-							/>
-							<p className="sm-text">
-								Once complete, NEO will be deposited to the address above
-							</p>
-						</div>
-						<div className="col-xs-4 center top-20">
-							{
-								isValidNeoOutput
-									? (<button onClick={this.handleOrderClick} className="btn-send">Place Oder</button>)
-									: (<div></div>)
-							}
-						</div>
-					</div>
-        
-					<div className="row">
-						<div className="col-xs-9 top-20">
-							<strong>
-              Minimum Order: 0.000000
-							</strong><br />
-							<span className="sm-text">Transaction fees included.</span>
-						</div>
-
-						<div className="col-xs-3">
-							<img
-								src={shapeshiftLogo}
-								alt=""
-								width="160"
-								className="logobounce"
-							/>
-						</div>
-
-					</div>
-				</div> )}
-
-			</div>
-		);
 	}
 }
